+++ libraries/prio-http/src/http2_client.cpp
@@ -90,7 +90,7 @@ ssize_t http2_client_stream::data_provider_callback(
     {                    
         *data_flags |= NGHTTP2_DATA_FLAG_EOF;
     }
-    return want;
+    return (ssize_t)want;
 }    
 
 /////////////////////////////////////////////////////
